+++ service-dbm/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "npx tsc",
-    "dev": "nodemon ./src/test.ts"
+    "dev": "nodemon ./src/main.ts"
   },
   "keywords": [],
   "author": "",
